Gloriana Class Ships & Mechanicum
Added placeholders for Gloriana Class ships (some have profiles, but many still need to be filled in - costs should be accurate, though). Added Mechanicum ships based on Battlefleet Heresy Harbingers of the Omnissiah. The Mechanicum rules are semi-complete - costs are accurate, but not all options are implemented yet. Added Scion of Prospero and Terminus Est as well.