Added Reserves catagory to gothic file
this wont be active for a while but I had to figure out how the profiles would need to be shared before building everything and having to change it later. Only core ship profiles should be in the gothic file to share. any options which they have added to them need to be done in the specific fleetlist/faction.